test(bot): split oversized queuemanipulation spec file
Split the monolithic 3248-line queueManipulation.spec.ts into 5 focused test files organized by concern: operations (shuffle, move, remove, rescue), deduplication logic, main replenish tests, advanced autoplay scenarios, and Spotify priority/diversity. Preserves all 79 tests with natural boundaries based on describe blocks and exported functions.

Closes #1618
